[python/r] Docstring audit for new shape
